:art: Refactored the remaining code :fire: Removed unused imports
