+++ src/main/java/com/okta/tools/OktaAwsCliAssumeRole.java
@@ -15,51 +15,23 @@
  */
 package com.okta.tools;
 
-import com.amazonaws.auth.AWSCredentialsProvider;
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.auth.BasicSessionCredentials;
-import com.amazonaws.regions.Regions;
-import com.amazonaws.services.securitytoken.AWSSecurityTokenService;
-import com.amazonaws.services.securitytoken.AWSSecurityTokenServiceClientBuilder;
 import com.amazonaws.services.securitytoken.model.AssumeRoleWithSAMLRequest;
 import com.amazonaws.services.securitytoken.model.AssumeRoleWithSAMLResult;
 import com.okta.tools.authentication.OktaAuthentication;
-import com.okta.tools.aws.settings.Configuration;
-import com.okta.tools.aws.settings.Credentials;
-import com.okta.tools.aws.settings.MultipleProfile;
 import com.okta.tools.aws.settings.Profile;
-import com.okta.tools.models.Session;
-import com.okta.tools.saml.*;
-import com.okta.tools.helpers.*;
-import org.apache.commons.lang.StringUtils;
-import org.apache.http.HttpStatus;
-import org.apache.http.StatusLine;
-import org.apache.http.client.entity.UrlEncodedFormEntity;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.entity.StringEntity;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.message.BasicNameValuePair;
+import com.okta.tools.helpers.ConfigHelper;
+import com.okta.tools.helpers.ProfileHelper;
+import com.okta.tools.helpers.RoleHelper;
+import com.okta.tools.helpers.SessionHelper;
+import com.okta.tools.saml.OktaSaml;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.select.Elements;
 
-import java.io.*;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.nio.file.Path;
+import java.io.IOException;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
-import java.util.*;
+import java.util.Optional;
+import java.util.Scanner;
 
 final class OktaAwsCliAssumeRole {
 
@@ -72,8 +44,8 @@ final class OktaAwsCliAssumeRole {
     private final String awsRoleToAssume;
     private final String oktaProfile;
 
-    static OktaAwsCliAssumeRole createOktaAwsCliAssumeRole(String oktaOrg, String oktaAWSAppURL, String oktaAWSUsername, String oktaAWSPassword,String oktaProfile, String oktaAWSRoleToAssume) {
-        return new OktaAwsCliAssumeRole(oktaOrg, oktaAWSAppURL, oktaAWSUsername, oktaAWSPassword,oktaProfile, oktaAWSRoleToAssume);
+    static OktaAwsCliAssumeRole createOktaAwsCliAssumeRole(String oktaOrg, String oktaAWSAppURL, String oktaAWSUsername, String oktaAWSPassword, String oktaProfile, String oktaAWSRoleToAssume) {
+        return new OktaAwsCliAssumeRole(oktaOrg, oktaAWSAppURL, oktaAWSUsername, oktaAWSPassword, oktaProfile, oktaAWSRoleToAssume);
     }
 
     private OktaAwsCliAssumeRole(String oktaOrg, String oktaAWSAppURL, String oktaUsername, String oktaAWSPassword, String oktaProfile, String awsRoleToAssume) {
@@ -86,30 +58,30 @@ final class OktaAwsCliAssumeRole {
     }
 
     String run(Instant startInstant) throws Exception {
-        Optional<com.okta.tools.models.Session> session = AwsSessionHelper.getCurrentSession();
-        Profile profile = AwsSessionHelper.getFromMultipleProfiles(oktaProfile);
+        Optional<com.okta.tools.models.Session> session = SessionHelper.getCurrentSession();
+        Profile profile = SessionHelper.getFromMultipleProfiles(oktaProfile);
 
-        if (session.isPresent() && AwsSessionHelper.sessionIsActive(startInstant, session.get()) && oktaProfile.isEmpty())
+        if (session.isPresent() && SessionHelper.sessionIsActive(startInstant, session.get()) && oktaProfile.isEmpty())
             return session.get().profileName;
 
         if (profile == null || startInstant.isAfter(profile.expiry)) {
             String oktaSessionToken = OktaAuthentication.getOktaSessionToken(getUsername(), getPassword(), oktaOrg);
             String samlResponse = OktaSaml.getSamlResponseForAws(oktaAWSAppURL, oktaSessionToken);
-            AssumeRoleWithSAMLRequest assumeRequest = chooseAwsRoleToAssume(samlResponse);
+            AssumeRoleWithSAMLRequest assumeRequest = RoleHelper.chooseAwsRoleToAssume(samlResponse, awsRoleToAssume);
             Instant sessionExpiry = startInstant.plus(assumeRequest.getDurationSeconds() - 30, ChronoUnit.SECONDS);
-            AssumeRoleWithSAMLResult assumeResult = assumeChosenAwsRole(assumeRequest);
-            String profileName = createAwsProfile(assumeResult);
+            AssumeRoleWithSAMLResult assumeResult = RoleHelper.assumeChosenAwsRole(assumeRequest);
+            String profileName = ProfileHelper.createAwsProfile(assumeResult, oktaProfile);
 
-            AwsConfigHelper.updateConfigFile(profileName, assumeRequest.getRoleArn());
-            AwsSessionHelper.addOrUpdateProfile(profileName, assumeRequest.getRoleArn(), sessionExpiry);
-            AwsSessionHelper.updateCurrentSession(sessionExpiry, profileName);
+            ConfigHelper.updateConfigFile(profileName, assumeRequest.getRoleArn());
+            SessionHelper.addOrUpdateProfile(profileName, assumeRequest.getRoleArn(), sessionExpiry);
+            SessionHelper.updateCurrentSession(sessionExpiry, profileName);
             return profileName;
         }
         return oktaProfile;
     }
 
     public void logoutSession() throws IOException {
-        AwsSessionHelper.logoutCurrentSession(oktaProfile);
+        SessionHelper.logoutCurrentSession(oktaProfile);
     }
 
     private String getUsername() {
@@ -137,130 +109,5 @@ final class OktaAwsCliAssumeRole {
         } else {
             return new String(System.console().readPassword("Password: "));
         }
-    }
-
-    private AssumeRoleWithSAMLResult assumeChosenAwsRole(AssumeRoleWithSAMLRequest assumeRequest) {
-        BasicAWSCredentials nullCredentials = new BasicAWSCredentials("", "");
-        AWSCredentialsProvider nullCredentialsProvider = new AWSStaticCredentialsProvider(nullCredentials);
-        AWSSecurityTokenService sts = AWSSecurityTokenServiceClientBuilder
-                .standard()
-                .withRegion(Regions.US_EAST_1)
-                .withCredentials(nullCredentialsProvider)
-                .build();
-        return sts.assumeRoleWithSAML(assumeRequest);
-    }
-
-    private AssumeRoleWithSAMLRequest chooseAwsRoleToAssume(String samlResponse) throws IOException {
-        Map<String, String> roleIdpPairs = AwsSamlRoleUtils.getRoles(samlResponse);
-        List<String> roleArns = new ArrayList<>();
-
-        String principalArn;
-        String roleArn;
-
-        if (roleIdpPairs.size() > 1) {
-            List<AccountOption> accountOptions = getAvailableRoles(samlResponse);
-
-            System.out.println("\nPlease choose the role you would like to assume: ");
-            System.out.println(roleIdpPairs.toString());
-            System.out.println(accountOptions.toString());
-            //Gather list of applicable AWS roles
-            int i = 0;
-            int j = -1;
-
-            for (AccountOption accountOption : accountOptions) {
-                System.out.println(accountOption.accountName);
-                for (RoleOption roleOption : accountOption.roleOptions) {
-                    roleArns.add(roleOption.roleArn);
-                    System.out.println("\t[ " + (i + 1) + " ]: " + roleOption.roleName);
-                    if (roleOption.roleArn.equals(awsRoleToAssume)) {
-                        j = i;
-                    }
-                    i++;
-                }
-            }
-            if ((awsRoleToAssume != null && !awsRoleToAssume.isEmpty()) && j == -1) {
-                System.out.println("No match for role " + awsRoleToAssume);
-            }
-
-            // Default to no selection
-            final int selection;
-
-            // If config.properties has matching role, use it and don't prompt user to select
-            if (j >= 0) {
-                selection = j;
-                System.out.println("Selected option "+ (j+1) + " based on OKTA_AWS_ROLE_TO_ASSUME value");
-            } else {
-                //Prompt user for role selection
-                selection = MenuHelper.promptForMenuSelection(roleArns.size());
-            }
-
-            roleArn = roleArns.get(selection);
-            principalArn = roleIdpPairs.get(roleArn);
-        } else {
-            Map.Entry<String, String> role = roleIdpPairs.entrySet().iterator().next();
-            System.out.println("Auto select role as only one is available : " + role.getKey());
-            roleArn = role.getKey();
-            principalArn = role.getValue();
-        }
-
-        return new AssumeRoleWithSAMLRequest()
-                .withPrincipalArn(principalArn)
-                .withRoleArn(roleArn)
-                .withSAMLAssertion(samlResponse)
-                .withDurationSeconds(3600);
-    }
-
-    private List<AccountOption> getAvailableRoles(String samlResponse) throws IOException {
-        Document document = getSigninPageDocument(samlResponse);
-        return AwsSamlSigninParser.parseAccountOptions(document);
-    }
-
-    private Document getSigninPageDocument(String samlResponse) throws IOException {
-        HttpPost httpPost = new HttpPost("https://signin.aws.amazon.com/saml");
-        UrlEncodedFormEntity samlForm = new UrlEncodedFormEntity(Arrays.asList(
-                new BasicNameValuePair("SAMLResponse", samlResponse),
-                new BasicNameValuePair("RelayState", "")
-        ), StandardCharsets.UTF_8);
-        httpPost.setEntity(samlForm);
-        try (CloseableHttpClient httpClient = HttpClients.createSystem();
-             CloseableHttpResponse samlSigninResponse = httpClient.execute(httpPost)) {
-            return Jsoup.parse(
-                    samlSigninResponse.getEntity().getContent(),
-                    StandardCharsets.UTF_8.name(),
-                    "https://signin.aws.amazon.com/saml"
-            );
-        }
-    }
-
-    private String createAwsProfile(AssumeRoleWithSAMLResult assumeResult) throws IOException {
-        BasicSessionCredentials temporaryCredentials =
-                new BasicSessionCredentials(
-                        assumeResult.getCredentials().getAccessKeyId(),
-                        assumeResult.getCredentials().getSecretAccessKey(),
-                        assumeResult.getCredentials().getSessionToken());
-
-        String awsAccessKey = temporaryCredentials.getAWSAccessKeyId();
-        String awsSecretKey = temporaryCredentials.getAWSSecretKey();
-        String awsSessionToken = temporaryCredentials.getSessionToken();
-
-        String credentialsProfileName = getProfileName(assumeResult);
-        AwsCredentialsHelper.updateCredentialsFile(credentialsProfileName, awsAccessKey, awsSecretKey, awsSessionToken);
-        return credentialsProfileName;
-    }
-
-    private String getProfileName(AssumeRoleWithSAMLResult assumeResult) {
-        String credentialsProfileName;
-        if(StringUtils.isNotBlank(oktaProfile))  {
-            credentialsProfileName = oktaProfile;
-        } else {
-            credentialsProfileName = assumeResult.getAssumedRoleUser().getArn();
-            if (credentialsProfileName.startsWith("arn:aws:sts::")) {
-                credentialsProfileName = credentialsProfileName.substring(13);
-            }
-            if (credentialsProfileName.contains(":assumed-role")) {
-                credentialsProfileName = credentialsProfileName.replaceAll(":assumed-role", "");
-            }
-        }
-        return credentialsProfileName;
     }
 }
