+++ src/main/java/com/okta/tools/helpers/ProfileHelper.java
@@ -1,0 +1,42 @@
+package com.okta.tools.helpers;
+
+import com.amazonaws.auth.BasicSessionCredentials;
+import com.amazonaws.services.securitytoken.model.AssumeRoleWithSAMLResult;
+import org.apache.commons.lang.StringUtils;
+
+import java.io.IOException;
+
+public class ProfileHelper {
+
+    public static String createAwsProfile(AssumeRoleWithSAMLResult assumeResult, String oktaProfile) throws IOException {
+        BasicSessionCredentials temporaryCredentials =
+                new BasicSessionCredentials(
+                        assumeResult.getCredentials().getAccessKeyId(),
+                        assumeResult.getCredentials().getSecretAccessKey(),
+                        assumeResult.getCredentials().getSessionToken());
+
+        String awsAccessKey = temporaryCredentials.getAWSAccessKeyId();
+        String awsSecretKey = temporaryCredentials.getAWSSecretKey();
+        String awsSessionToken = temporaryCredentials.getSessionToken();
+
+        String credentialsProfileName = getProfileName(assumeResult, oktaProfile);
+        CredentialsHelper.updateCredentialsFile(credentialsProfileName, awsAccessKey, awsSecretKey, awsSessionToken);
+        return credentialsProfileName;
+    }
+
+    private static String getProfileName(AssumeRoleWithSAMLResult assumeResult, String oktaProfile) {
+        String credentialsProfileName;
+        if (StringUtils.isNotBlank(oktaProfile)) {
+            credentialsProfileName = oktaProfile;
+        } else {
+            credentialsProfileName = assumeResult.getAssumedRoleUser().getArn();
+            if (credentialsProfileName.startsWith("arn:aws:sts::")) {
+                credentialsProfileName = credentialsProfileName.substring(13);
+            }
+            if (credentialsProfileName.contains(":assumed-role")) {
+                credentialsProfileName = credentialsProfileName.replaceAll(":assumed-role", "");
+            }
+        }
+        return credentialsProfileName;
+    }
+}
