+++ src/main/java/com/okta/tools/models/RoleOption.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2017 Okta
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.okta.tools.models;
+
+import java.util.Objects;
+
+public final class RoleOption {
+    public final String roleName;
+    public final String roleArn;
+
+    public RoleOption(String roleName, String roleArn) {
+        this.roleName = roleName;
+        this.roleArn = roleArn;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        RoleOption that = (RoleOption) o;
+        return Objects.equals(roleName, that.roleName) &&
+                Objects.equals(roleArn, that.roleArn);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(roleName, roleArn);
+    }
+
+    @Override
+    public String toString() {
+        return "RoleOption{" +
+                "roleName='" + roleName + '\'' +
+                ", roleArn='" + roleArn + '\'' +
+                '}';
+    }
+}
